+++ assets/js/selfoss-events-entriestoolbar.js
@@ -56,10 +56,10 @@ selfoss.events.entriesToolbar = function(parent) {
         if (parent.find('.entry-toolbar').has('button.entry-share' + shares[0]).length == 0) {
             // add the share toolbar entries
             parent.find('.entry-smartphone-share button.entry-newwindow').parent().after(selfoss.shares.buildLinks(shares, function(name) {
-                return '<li><button type="button" class="entry-share entry-share' + name + '" title="' + name + '" aria-label="' + name + '"><img class="entry-share" title="' + name + '" alt="' + name + '" src="images/' + name + '.png" height="16" width="16">' + name + '</button></li>';
+                return '<li><button type="button" class="entry-share entry-share' + name + '" title="' + name + '" aria-label="' + name + '">' + selfoss.shares.icons[name] + ' ' + name + '</button></li>';
             }));
             parent.find('.entry-toolbar button.entry-next').parent().after(selfoss.shares.buildLinks(shares, function(name) {
-                return '<li><button type="button" class="entry-share entry-share' + name + '" aria-label="' + name + '"><img title="' + name + '" alt="' + name + '" src="images/' + name + '.png" height="16" width="16"></button></li>';
+                return '<li><button type="button" class="entry-share entry-share' + name + '" title="' + name + '" aria-label="' + name + '">' + selfoss.shares.icons[name] + '</button></li>';
             }));
             // hookup the share icon click events
             for (var i = 0; i < shares.length; i++) {